app/requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# File uploads
python-multipart==0.0.9

# Google Gemini (Generative AI SDK)
google-generativeai==0.8.3

# Text-to-Speech (Coqui TTS)
TTS==0.22.0

# Environment variables
python-dotenv==1.0.1
